Don't fail the autopkgtest job if all tests are skipped or superficial.
autopkgtest returns exit code 8 if there no tests in the package, or if all
non-superficial tests were skipped, or if all tests pass but are marked as
"superficial" (for details about the "superficial" test restriction, see
https://bugs.debian.org/904979). Whenever this happens, on the current Debian
infrastructure the test results won't affect migration-to-testing delay: they
won't add delay as a failure would, but they won't decrease the delay either as
successful non-superficial tests would.

Assuming there's no way to express this in GitLab CI (AFAIK a test can only
succeed or fail), in this situation it's better to succeed than to fail: if we
always consider such skipped/superficial test suite runs as failed, the package
maintainer will simply get used to seeing them in failed state and will not have
any easy mean to notice if a regression makes the autopkgtests actually fail…
which kinda defeats the point of running them in the first place.