move --no-diffoscope a default argument for reprotest
Big projects can fail due to insufficient memory to run reprotest with
diffoscope, but small projects may want to enable it as an easy way to
locate problems. This change introduces a new REPROTEST_EXTRA_ARGS that
has default value '--no-diffoscope' in reprotest template, so a derived
job may override it with:

  variables:
    REPROTEST_EXTRA_ARGS: ''

A test-reprotest-diffoscope job template is also added to serve with
diffoscope enabled by default.

Close: #17