Add additional jobs for arch=all and arch=any
Testing if it is possible to build only the arch=all or arch=any
packages from a source package is important because failures here can
lead to FTBFS on the buildds for architectures other than the one a DD
uploaded as or during binNMUs or when doing source-only uploads.

Unfortunately the pipelines can't detect the what kind of binary package
will be created so the jobs added have to be selected manually by the
interested user.

But providing those additional targets is a start, the mechanism can be
improved in the future.

The new jobs don't contribute artifacts to the whole pipeline to not
confuse the depending jobs. Only the normal full builds artifacts are
used for the later stages.