Enhance the auto-selection of build variants to do
Reduce the magic of the if-statement checking if a build job should run.
Allow it to not only match "Architecture: any" but also include the
local build architecture in the match, to also run the arch-specific
build when we are dealing with a package with a specific set of
architectures instead of just "any".