Make autopkgtest run like in ci.d.n \o/
New autopkgtest image includes lxc. An lxc image is generated on
salsa-ci-pipeline/autopkgtest-lxc to save time and is dowloaded here.

As in ci.d.n, autopkgtest is always the last version and the flavours
are changed on the lxc testbed, so from now on we can remove all the
autopkgtest:{stretch, jessie, etc} and just build for latest .