Improve piuparts test
Those changes are required to make the test pass on Kali. There's
some logic about alternate distribution which is specific to Debian
and that we include only for Debian.

Also piuparts doesn't know about Kali so it must be explicitly given the
default release and some extra options.

The changes here should not have any effect on the behaviour of piuparts
in the Debian case. On contrary, it should help piuparts to adjust its
behaviour if we are running tests for a release which is not unstable.