Improve the license and contributor checks
1/ You want to run both checks in all cases so that contributors see both
errors on the first try, otherwise they will fix the license problem
first, send back their MR and have another fail because they missed to
add themselves to the CONTRIBUTORS file.

2/ You want to print exactly which are the problematic files and you
want to indicate the user identity that you are looking for in the
contributors file.

3/ A contributor might not want to use the email associated to his
gitlab account so we also offer the possibility to pass the check if the
real name appears in the file, even if it's with another email.