Add a missing_breaks test
This test checks if the files provided by the built packages conflict
with another package without having the corresponding
Breaks/Replaces/Conflicts, and produces the missing_breaks.xml junit
file with the result.

A new tests image is added as a generic image for this kind of "smaller"
tests, and thus avoiding maintaining a tests for each RELEASE for each
test.