Add support for debian only packaging branches
Teams like KDE use debian only packaging branches, and rely on origtargz
to get the uploaded tarball from the archive, and the new versions using
uscan.

This commit adds some logic regarding adding the deb-src entries
(needed for origtargz), and splits the gbp buildpackage in a export-orig
and and a buildpackage, calling origtargz when the export-orig fails.