Allow passing extra apt repositories and keys
This adds two new user configurable variables that make it possible to
build packages with build dependencies that are not currently in the
archive.

Both EXTRA_REPOSITORY and EXTRA_REPOSITORY_KEY are of type file, this
simplifies handling multiple lines.

The add_extra_repository.sh script copies the contents of the files to
the corresponding locations. This script needs to be present in all
images that require installing updated dependencies. And gets called
from the job definitions.