Ignore add_extra_repository.sh errors
The pipeline tests currently use the published docker images, thus
adding a script with the dockerfiles has no effect on the pipelines test
till the change is approved, and a change that adds a script in the
dockerfiles to call it from the pipelines test job is doomed to fail
(See !139).

This change simply ignores the missing script in the pipeline test.