Delete bogus '$CI_PROJECT_DIR' dir
The report #104, followed by the mr !163 documents the issue but the
applied fix tries to delete the '$CCACHE_WORK_DIR' '${CCACHE_WORK_DIR}'
dirs, which as the variables are not replaced doesn't fix the problem as
intended. This commits fixes that by doing what's documented.

See https://salsa.debian.org/cinnamon-team/python-xapp/-/jobs/288475 for
an example of an affected job with the !163 applied.