Do not run the pipeline for tags by default
Currently, if pushing a commit and a corresponding tag, the pipeline
will run two times. In the huge majority of cases, both pipelines will
have the same result, which seems like a waste of resources, and
produces quite some noise, specifically if teams are using notifications
sent to IRC or via mail.

This commit tries to improve this situation via disabling the pipeline
for tags, by default. However, it's still possible to run the pipeline
if tags are pushed, via setting the variable
SALSA_CI_ENABLE_PIPELINE_ON_TAGS to one of "1", "yes" or "true".

Closes #39
Closes #77

ok