Add unzip to the gbp docker image
This is needed when origtargz downloads the current upstream version as
a .zip archive.

    ~~ Detailed explanations ~~

The gbp docker image is used for the job '.provisioning-extract-source'.

Looking into the job definition in the file 'salsa-ci.yaml', we can see
that the job might run 'origtargz',  which might end up running 'uscan
--download --download-current-version' (see origtargz(1) for details of
what it does).

Uscan downloads the upstream tarball and unpacks it. If ever the
upstream tarball is a zip file, 'unzip' is required to unpack it.

To demonstrate the failure, I forked an existing package, and then:
1) changed d/watch to use zip
2) bumped d/changelog to the latest upstream version
3) git push and let the pipeline run

This can be see at:
1) https://salsa.debian.org/arnaudr/goodvibes/-/commit/84f7920
2) https://salsa.debian.org/arnaudr/goodvibes/-/commit/a5ce2b2
3) https://salsa.debian.org/arnaudr/goodvibes/-/pipelines/256210
   -> https://salsa.debian.org/arnaudr/goodvibes/-/jobs/1656082

The actual error being:

  Trying uscan --download --download-current-version ...
  uscan: Newest version of goodvibes on remote site is 0.6.3, specified download version is 0.6.3
  uscan error: unzip binary not found. You need to install the package unzip to be able to repack zip upstream archives.