README.md: Clarify that allow_failure applies to any job, not only reprotest
While reading the README, it seemed to me that allow_failure was a
setting specific to the Salsa CI Pipeline, and even more, specific to
the reprotest job. Only later I realized that it was a GitLab CI
setting, and that it could be applied to pretty much any job.

This commit clarifies this point a bit, while still keeping reprotest as
an example of a job that can be allowed to fail.