Set RELEASE=kali-dev (instead of sid) in .gitlab-ci.yml
For Kali, we don't want the release to be `sid`, but `kali-dev`.

For the `test-pipeline` job, it seems that we could just drop the
variable assignment, as this job includes `.pipeline-test.yml`, which
itself includes `vars/kali.yml`, and in there the variable `RELEASE` is
set to `kali-dev`.

For the `check license + contributor` job though, it seems that we must
set it expicitly.

No idea about the `clean images` job, but for details refer to
09c3ca78bcca4ec9ca078ee819460aece244d7d8