Bump dind image for the 'clean images' job
This is just a follow-up of 19f85f8b7f9e22ad8f12bc5d844422e070b8b4ee.
The idea is to use the same version of the dind image everywhere, for
simplicity.

This change doesn't fix anything, and it doesn't break anything either,
to the best of my knowledge. This change has been deployed in Kali Linux
CI for a few weeks and nothing bad happened so far.