README: Add a TL;DR section
I have to go to the README file each time I'm configuring a salsa repo,
but the string I'm looking for is hidden deep inside the text. Show it
at the front so it's visible.

Also update the title used in the Gitlab config, the setting is now
called "CI/CD configuration file", and add "variables:" as context for
SALSA_CI_BUILD_TIMEOUT_ARGS.

Signed-off-by: Christoph Berg <myon@debian.org>