images/dockerfiles/piuparts.0: install usrmerge
It seems the docker images are not merged-usr by default. Install usrmerge
in the piuparts image, otherwise when usrmerge is made pseudo-essential it
will be pulled in at test time, modifying /etc/shells and causing piuparts
to flag it as a package-under-test problem, when it really isn't.

E.g.: https://salsa.debian.org/debian/init-system-helpers/-/jobs/3241271