Add support for Kali Linux base images
In the Docker Hub, only the "Official Docker Images" have names such as
ARCH/IMAGE. Images provided by not-Docker don't have access to the ARCH
namespace.

Therefore with this commit, we move the code that combines ARCH and
IMAGE in a bash snippet in .images-ci.yml, and we gain enough
flexibility to support images from other distributions.

We can then add support for Kali Linux images.