Provide Lintian images for all releases
Since !347 the Lintian job runs by default with the Sid
version of Lintian irrespective of what ${RELEASE} is
used. If a Debian Developer wants to do quality assurance
to a package in an existing release (e.g Bullseye) they
can't run the Sid Lintian (currently v2.115) as it will
throw way too many errors and warnings than what is
reasonable to expect to be fixed in a stable release.

Thus for a Bullseye maintenance it would be better to
continue using the Bullseye version of Lintian and focus
only on ensuring there are no new errors/warnings compared
to what there was previously for the package in Bullseye.

Developers could define in their salsa-ci.yml:

  RELEASE: bullseye
  SALSA_CI_IMAGES_LINTIAN: ${SALSA_CI_IMAGES}/lintian:bullseye

However, for this to work also the actual Lintian image
for Bullseye must exist, so change the build matrix to
build Lintian images for all releases.