Don't unset HOST_ARCH in build-before-script, not needed
It doesn't seem to be needed.

The variable HOST_ARCH is empty by default. The only job that sets it is
the job test-crossbuild-package-arm64. However, this job also has the
following rule:

    rules:
      - if: $BUILD_ARCH == $HOST_ARCH
        when: never

So we already ensure that, when HOST_ARCH is set, it's also different
from BUILD_ARCH. There's no need to do the test a second time.

This commit is in preparation for the next commit.