Move the "d/control's Architecture test" before build-before-script
Looking at what build-before-script does, it's completely unrelated to
checking the Architecture field in debian/control, so it doesn't matter
if we do it before or after.

The only side-effect *was* that HOST_ARCH might have been modified in
build-before-script, but I just removed that in previous commit. So now
it's safe to swap these two steps.

The main gain from this move is simplification: in case we wan't to exit
early, we don't need to move CCACHE_TMP_DIR back to CCACHE_WORK_DIR,
since build-before-script (which moved it in the other direction) didn't
run yet.

The second gain is that it will enable more code refactor in the
upcoming commits.