Fix test-(cross)?build-package-* for .pipeline-test
The 3 jobs test-(cross)?build-package-* are actually broken when they
are run via .pipeline-test.yml. It's not obvious at all, since the jobs
pass (they are green), but looking at the logs (for example,
https://salsa.debian.org/salsa-ci-team/pipeline/-/jobs/4376399), we can
see those lines:

    grep: debian/control: No such file or directory
    ###########################################
    ### No binary package of type all found
    ### Stopping test-build-all test.
    ###########################################
    You should disable this job via:
    variables:
      SALSA_CI_DISABLE_BUILD_PACKAGE_ALL: '1'

Why is debian/control missing? Because the check is done at a moment
when we are in the git checkout, and not in the "debianized source
package".

To clarify: this is Ok for the vast majority of the packages that run
the pipeline, because for those, the git checkout is almost equal to the
debianized source package, and debian/control is there, in the git
checkout.

However, salsa-ci-pipeline is "special": .pipeline-test.yml doesn't run
the pipeline on itself, it runs it for another package (grep), cf.
- .gitlab.yml 's test-pipeline
- .pipeline-test.yml 's extract-source

So let's fix it. Let's make sure to the "debianized source package"
(that is created by the extract-source job) first, before doing any
checks against the source package. Then move back afterward.

It's a bit ugly that we have to do that in 3 jobs, but hopefully those
jobs will be merged into one at some point (they are almost identical).