SALSA_CI_DISABLE_ALL_TESTS & ..._ENABLE_<TEST>
SALSA_CI_DISABLE_ALL_TESTS:
  disables all the test jobs in the salsa-CI pipeline.

This can be overriden by asserting `SALSA_CI__ENABLE_<TEST>` for any
`<TEST>` that is required.

The intent here is that if an individual test is both DISABLED (either
directly or via the _ALL_TESTS flag) and ENABLED, it should act as
though it had not been disabled.