aptly job: handle multiple artifact dirs
This allows me to add a job which harvests artifacts from various repos
and drops them in e.g. ${WORKING_DIR}_harvest, and then to define a
variant of aptly which needs: the artifacts from that job, with the
result that one gets an aptly repo that contains the combination of all
the (u)debs from all the projects in question.

This means that most of branch2repo can be discarded in favour of a
slightly extended aptly job definition.