Create job .test-build-package
This job is quite similar to the job .build-package, which allows to
factorize code between all the build-package-* jobs.

In the same spirit, we define .test-build-package to factorize code
between all the test-build-package-* and test-crossbuild-* jobs.

This commit only defines this new job, and the next commit will make use
of it.