Disable test-package-build-twice for Kali
This new test is enabled by default in Salsa CI, per Debian Policy, cf.
https://salsa.debian.org/salsa-ci-team/pipeline/-/merge_requests/438#note_422145

However for Kali, it doesn't really matter, there's no such policy, and
since we build all our package with gbp, what matters is the Git repo.

So this test has only downsides for Kali: it will catch issues that we
don't care about and don't plan to fix, it will make pipelines longer to
run, it might even fail packages for which the build duration is already
close to the max job duration limit (I suppose that the run-twice job
will take 2 times more to run that a normal build).

Hence we disable this job for Kali.